requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pydantic>=2.0
joblib
pyyaml
pytest
mlflow